Update readme to set --max-chain-length
This also includes a number of other adjustments to required location
flags.

I have also tried to make it clearer when the google project must be set
in the command without adding too much noise to the README. Please feel
free to suggest improvements here, we might just set and assume PROJECT
is set throughout?

Signed-off-by: Charlie Egan <charlieegan3@users.noreply.github.com>